Move PluginLoader to a separate class
